requirements.txt
openai
langchain
langchain-community
langchain-openai
chromadb
tiktoken
python-dotenv
streamlit
